Add support for geo-intents.
Allows both setting markers for that position as well
as translating the given position.
Very useful both for testing as well as populating
initial markers without having to walk around a lot,
e.g. if Osmand contains only a select few of the features
in the map image you have.

Signed-off-by: Reimar Döffinger <Reimar.Doeffinger@gmx.de>